Napraw miganie podgladu i rozmieszczenie okien; dodaj ustawienia
- FolderWatcherService/PhotoLibraryService: pliki obecne w folderze
  przy starcie nie wywoluja juz NewestChanged pojedynczo - usuwa to
  gwaltowne przewijanie zdjec w Oknie A przy starcie z pelnym folderem.
- Okna sa teraz pozycjonowane przez Win32 SetWindowPos w pikselach
  fizycznych zamiast WPF Left/Top/Width/Height w DIP - naprawia zle
  rozmieszczenie na monitorach ze skalowaniem innym niz 100%, przez
  ktore siatka (Okno B) bywala calkowicie przykryta przez Okno A.
- Nowe okno ustawien (SettingsWindow): SSID/haslo WiFi dla gosci,
  liczba kolumn siatki, czas podgladu wybranego zdjecia - pytanie
  o zmiane ustawien pojawia sie po wyborze folderu przy starcie.
- Zakladka "O aplikacji" w oknie ustawien: autor + link do repo.

+++ src/Fotopodglad/Models/ScreenInfo.cs
@@ -1,12 +1,18 @@
 namespace Fotopodglad.Models;
 
-/// <summary>Geometria monitora w jednostkach WPF (DIP), już przeliczona z pikseli fizycznych przez współczynnik DPI danego ekranu.</summary>
+/// <summary>
+/// Geometria monitora w PIKSELACH FIZYCZNYCH (surowe Screen.Bounds), celowo bez żadnego przeliczania
+/// przez DPI. Okna są pozycjonowane przez Win32 SetWindowPos w tych samych jednostkach (zob.
+/// App.ConfigureFullscreenWindow) — to omija cały problem niejednoznaczności DIP-na-monitor-o-innym-DPI,
+/// który przy oknach ustawianych przez WPF Left/Top/Width/Height (jednostki DIP) prowadził do złego
+/// rozmiaru/pozycji okna na monitorach ze skalowaniem innym niż 100%.
+/// </summary>
 public sealed class ScreenInfo
 {
     public required string DeviceName { get; init; }
-    public required double Left { get; init; }
-    public required double Top { get; init; }
-    public required double Width { get; init; }
-    public required double Height { get; init; }
+    public required int Left { get; init; }
+    public required int Top { get; init; }
+    public required int Width { get; init; }
+    public required int Height { get; init; }
     public required bool IsPrimary { get; init; }
 }
